requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-multipart
psycopg2-binary
pydantic[email]
pytest
httpx
alembic
sqlalchemy